[ANDROAPP-3890] FormEventsRepository derived
