+++ app/src/main/java/org/dhis2/usescases/eventsWithoutRegistration/eventCapture/EventCaptureModule.java
@@ -17,10 +17,11 @@ import org.dhis2.data.forms.dataentry.ValueStore;
 import org.dhis2.data.forms.dataentry.ValueStoreImpl;
 import org.dhis2.data.forms.dataentry.fields.FieldViewModelFactory;
 import org.dhis2.data.forms.dataentry.fields.FieldViewModelFactoryImpl;
-import org.dhis2.form.data.FormRepositoryImpl;
-import org.dhis2.form.model.RowAction;
 import org.dhis2.data.prefs.PreferenceProvider;
 import org.dhis2.data.schedulers.SchedulerProvider;
+import org.dhis2.form.data.FormEventsRepository;
+import org.dhis2.form.data.FormRepositoryImpl;
+import org.dhis2.form.model.RowAction;
 import org.dhis2.utils.RulesUtilsProvider;
 import org.dhis2.utils.resources.ResourceManager;
 import org.hisp.dhis.android.core.D2;
@@ -119,7 +120,11 @@ public class EventCaptureModule {
 
     @Provides
     @PerActivity
-    org.dhis2.form.data.FormRepository  provideFormRepository(@NonNull D2 d2) {
-        return new FormRepositoryImpl(d2, eventUid);
+    org.dhis2.form.data.FormRepository provideEventsFormRepository(@NonNull D2 d2) {
+        return new FormEventsRepository(
+                new FormRepositoryImpl(),
+                d2,
+                eventUid
+        );
     }
 }
