+++ form/src/test/java/org/dhis2/form/data/FormEventsRepositoryTest.kt
@@ -1,0 +1,10 @@
+package org.dhis2.form.data
+
+import org.hisp.dhis.android.core.D2
+import org.mockito.Mockito
+
+class FormEventsRepositoryTest {
+
+    private val d2: D2 = Mockito.mock(D2::class.java, Mockito.RETURNS_DEEP_STUBS)
+    private val recordUid = "ABCD"
+}
